% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{Determine Distance Between Two Points}
\usage{
distance(long1, lat1, long2, lat2)
}
\arguments{
\item{long1}{The longitude for point one.}

\item{lat1}{The latitude for point one.}

\item{long2}{The longitude for point two.}

\item{lat2}{The latitude for point two.}
}
\description{
Calculate the distance between two points based on their latitude and
longitude.
}
