% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapBubbles.R
\name{mapBubbles}
\alias{mapBubbles}
\title{Create a Map}
\usage{
mapBubbles(
  dat,
  depths = NULL,
  plotMap = T,
  maxValue = max(dat$Z),
  maxSize = 0.3,
  circle.fg = "black",
  circle.bg = grDevices::rgb(0, 0, 0, 0.3),
  type = "surface",
  zeroPch = 13,
  zeroCex = 0.8,
  zeroCol = grDevices::gray(0.5),
  legendPos = "bottomright",
  legendBreaks = NULL,
  legendType = "vert",
  legendTitle = NULL,
  legendCex = 0.8,
  legendTitleCex = legendCex,
  ...
)
}
\arguments{
\item{dat}{The data.}

\item{depths}{The default is \code{NULL}.}

\item{plotMap}{A logical value specifying if the map should be plotted.}

\item{maxValue}{todo: document the argument.}

\item{maxSize}{todo: document the argument.}

\item{circle.fg}{The color for the foreground of the bubles.}

\item{circle.bg}{The color for the background of the bubles.}

\item{type}{todo: document the argument.}

\item{zeroPch}{todo: document the argument.}

\item{zeroCex}{todo: document the argument.}

\item{zeroCol}{todo: document the argument.}

\item{legendPos}{todo: document the argument.}

\item{legendBreaks}{todo: document the argument.}

\item{legendType}{todo: document the argument.}

\item{legendTitle}{todo: document the argument.}

\item{legendCex}{todo: document the argument.}

\item{legendTitleCex}{todo: document the argument.}

\item{...}{Arguments passed to \code{plotMap}.}
}
\description{
Produce a map with bubbles for points.
}
