% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiMap.R
\name{multiMap}
\alias{multiMap}
\title{Make Map}
\usage{
multiMap(nrows = 1, ncols = 1, lonLims, latLims, adjLonRange = TRUE, fnc, ...)
}
\arguments{
\item{nrows}{The number of rows.}

\item{ncols}{The number of columns.}

\item{lonLims}{The limits for Longitude.}

\item{latLims}{The limits for Latitude.}

\item{adjLonRange}{A logial value.}

\item{fnc}{todo: document fnc argument.}

\item{...}{Arguments passed to fnc.}
}
\description{
todo: better document description of multiMap
}
