\name{nwfscMapping-package}
\alias{nwfscMapping-package}
\alias{nwfscMapping}
\docType{package}
\title{Mapping tools for NWFSC survey data}
\description{
Some mapping tools for NWFSC assessments
}
\details{
    \tabular{ll}{
        Package: \tab nwfscMapping\cr
        Type: \tab Package\cr
        Version: \tab 1.0\cr
        Date: \tab 2013-01-01\cr
        License: \tab What license is it under?\cr
    }
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
    Allan C Hicks <allan.hicks_at_noaa.gov>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
