\name{WCstatesInlandPBS}
\alias{WCstatesInlandPBS}
\docType{data}
\title{
West Coast state boundaries
}
\description{
West Coast state boundaries
}
\usage{data(WCstatesInlandPBS)}
\format{
  A data frame with 106 observations on the following 5 variables.
  \describe{
    \item{\code{PID}}{a numeric vector}
    \item{\code{SID}}{a numeric vector}
    \item{\code{POS}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
  }
}
\examples{
data(WCstatesInlandPBS)
## maybe str(WCstatesInlandPBS) ; plot(WCstatesInlandPBS) ...
}
\keyword{datasets}
