\name{westCoastLL}
\alias{westCoastLL}
\docType{data}
\title{
    West Coast Lat/Lon ...
}
\description{
Latitude Longitude of the West Coast
}
\usage{data(westCoastLL)}
\format{
  Latitude Longitude of the West Coast stitched together from nepacLL in PBSmapping and worldHiRes
  \describe{
    \item{\code{PID}}{a numeric vector}
    \item{\code{POS}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
  }
}
\examples{
data(westCoastLL)
## maybe str(westCoastLL) ; plot(westCoastLL) ...
}
\keyword{datasets}
